apps/dash-label-properties/requirements.txt
dash
dash-core-components
dash-html-components
dash-canvas>=0.0.8
dash-table
plotly
gunicorn
scikit-image
scikit-learn
pandas
numpy
Pillow
dash-bootstrap-components==0.10.7